renames graphs in explorer
