+++ src/components/explore/AggregationComparisonGraph.svelte
